Reorganize `pyramid.path`.
Move resolver-related things into `pyramid.resolver`.  Move the rest
into `pyramid._path`, which is necessary to resolve circular for future
changes.
